test(package-manager): extract inline test modules per #369 convention
main landed #369 ("extract all inline test modules into dedicated
files") while this PR was open, switching the workspace convention
to external `tests.rs` files for every test module. The unit tests
this PR added in `create_virtual_dir_by_snapshot.rs`,
`install_package_by_snapshot.rs`, and `create_virtual_store.rs`
were inline; move each one to its dedicated `<parent>/tests.rs`
with a `#[cfg(test)] mod tests;` declaration on the parent, matching
the rest of the package-manager crate.

No behavioral change.
